+++ src/aui-form-builder/js/aui-form-builder-field-choice.js
@@ -11,6 +11,7 @@ var CSS_FIELD_CHOICE = A.getClassName('form', 'builder', 'field', 'choice'),
     CSS_FIELD_CHOICE_OPTION_OTHER = A.getClassName('form', 'builder', 'field', 'choice', 'option', 'other'),
     CSS_FIELD_CHOICE_OPTIONS = A.getClassName('form', 'builder', 'field', 'choice', 'options'),
     CSS_FIELD_CHOICE_TITLE = A.getClassName('form', 'builder', 'field', 'choice', 'title'),
+    CSS_FIELD_CONTENT = A.getClassName('form', 'builder', 'field', 'content'),
 
     TPL_FIELD_CHOICE_CONTENT = '<div>' +
         '<label class="' + CSS_FIELD_CHOICE_TITLE + '"></label>' +
@@ -39,7 +40,7 @@ A.FormBuilderFieldChoice = A.Base.create('form-builder-field-choice', A.FormBuil
         var content = this.get('content');
 
         content.addClass(CSS_FIELD_CHOICE);
-        content.one('.form-builder-field-content').setHTML(TPL_FIELD_CHOICE_CONTENT);
+        content.one('.' + CSS_FIELD_CONTENT).setHTML(TPL_FIELD_CHOICE_CONTENT);
 
         this._uiSetHelp(this.get('help'));
         this._uiSetTitle(this.get('title'));